cleanup and restructure for rebuild
