+++ .gitignore
@@ -2,12 +2,10 @@
 *.db*
 *.log*
 *.info
-rh.json
 venv
 settings.py
 .DS_Store
 macos.sh
 .vscode
-scanner.py
 src/db.json
 node_modules
