--- Aptfile
@@ -1,13 +1,0 @@
-g++
-gcc
-golang-go
-libc6-dev
-make
-pkg-config
-wget
-tmux
-python3.6
-python3-pip
-python3-setuptools
-python3-dev
-git
