--- Pipfile
@@ -1,73 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-
-[packages]
-attrs = "==19.3.0"
-"backports.functools-lru-cache" = "==1.6.1"
-beautifulsoup4 = "==4.9.0"
-certifi = "==2020.4.5.1"
-chardet = "==3.0.4"
-click = "==7.1.2"
-cobe = "==3.0.0"
-configparser = "==5.0.0"
-contextlib2 = "==0.6.0.post1"
-idna = "==2.9"
-importlib-metadata = "==1.6.0"
-importlib-resources = "==1.5.0"
-inflect = "==4.1.0"
-irc = "==18.0.0"
-itsdangerous = "==1.1.0"
-"jaraco.classes" = "==3.1.0"
-"jaraco.collections" = "==3.0.0"
-"jaraco.functools" = "==3.0.1"
-"jaraco.itertools" = "==5.0.0"
-"jaraco.logging" = "==3.0.0"
-"jaraco.stream" = "==3.0.0"
-"jaraco.text" = "==3.2.0"
-joblib = "==0.14.1"
-memory-profiler = "==0.57.0"
-more-itertools = "==8.2.0"
-mypy = "==0.770"
-mypy-extensions = "==0.4.3"
-nltk = "==3.5"
-pathlib2 = "==2.3.5"
-pluggy = "==0.13.1"
-praw = "==7.0.0"
-prawcore = "==1.3.0"
-psutil = "==5.7.0"
-py = "==1.8.1"
-pyparsing = "==2.4.7"
-pytest = "==5.4.1"
-pytest-mypy = "==0.4.2"
-pytz = "==2020.1"
-regex = "==2020.4.4"
-requests = "==2.23.0"
-scandir = "==1.10.0"
-singledispatch = "==3.4.0.3"
-six = "==1.14.0"
-soupsieve = "==2.0"
-tempora = "==3.0.0"
-tinydb = "==4.0.0.post1"
-tqdm = "==4.46.0"
-typed-ast = "==1.4.1"
-typing = "==3.7.4.1"
-typing-extensions = "==3.7.4.2"
-update-checker = "==0.17"
-urllib3 = "==1.25.9"
-wcwidth = "==0.1.9"
-zipp = "==3.1.0"
-Cython = "==0.29.17"
-Flask = "==1.1.2"
-Jinja2 = "==2.11.2"
-MarkupSafe = "==1.1.1"
-PyStemmer = "==2.0.0.1"
-websocket_client = "==0.57.0"
-Werkzeug = "==1.0.1"
-imgurpython = "*"
-
-[requires]
-python_version = "3.7"
