--- Procfile
@@ -1,1 +1,0 @@
-worker: python3 ./src/run.py
