--- src/disallowed_subs.txt
@@ -1,108 +1,0 @@
-abuse
-abusiverelationships
-anxiety
-anxietyhelp
-bipolar
-confession
-deadredditors
-depressed
-depression
-depression_help
-depression_memes
-depressionandPTSD
-foreveralone
-introvert
-kindvoice
-lonely
-mentalhealth
-mentalillness
-ocd
-ptsd
-raisedbynarcissists
-relationship_advice
-relationships
-sad
-selfHarmScars
-selfharm
-selfhelp
-socialanxiety
-stopselfharm
-suicidewatch
-therapy
-depressionregimens
-adhd
-drugs
-eatingdisorders
-schizophrenia
-borderlinepdisorder
-cptsd
-bipolar2
-stopdrinking
-benzorecovery
-quittingkratom
-addiction 
-addictionprevention 
-alcoholism
-stopsmoking
-petioles
-stopspeeding
-healthproject
-rapecounseling
-aspergers
-behaviortherapy
-bipolarreddit
-bodyacceptance 
-downsyndrome
-dysmorphicdisorder 
-healthanxiety
-ocpd
-shizoaffective
-tourettes
-survivorsofabuse
-rape
-inthenews
-usnews
-thenews
-anythinggoesnews
-qualitynews
-fullnews
-full_news
-neutralnews
-worldnews
-news
-offbeat 
-gamernews 
-upliftingnews
-politics
-wordpolitics
-libertarian
-anarchism 
-socialism 
-animetitties
-conservative
-ukpolitics
-geopolitics
-communism
-completeanarchy 
-chapotraphouse 
-mensrights
-feminism 
-sandersforpresident 
-bidenpro 
-political_revolution
-thanksobama
-esist 
-the_mueller
-fuckthealtright
-democrats
-liberal
-keep_track
-enoughtrumpspam
-marchagainsttrump
-tinytrumps
-trumpcriticizestrump 
-trumpgret
-impeach_trump 
-energy
-vent
-twoxchromosomes
