--- src/requirements.old
@@ -1,27 +1,0 @@
-backports.functools-lru-cache==1.4
-beautifulsoup4==4.7.1
-certifi==2017.4.17
-chardet==3.0.4
-cobe==2.1.2
-idna==2.5
-inflect==0.2.5
-irc==12.1.1
-jaraco.classes==1.4.1
-jaraco.collections==1.5.1
-jaraco.functools==1.16
-jaraco.itertools==2.0.1
-jaraco.logging==1.5
-jaraco.text==1.9.2
-more-itertools==3.2.0
-nltk==3.4
-praw==6.0.0
-prawcore==1.0.0
-PyStemmer==1.3.0
-pytz==2017.2
-requests>=2.20.0
-singledispatch==3.4.0.3
-six==1.10.0
-tempora==1.8
-update-checker==0.16
-urllib3>=1.23
-tinydb==3.15.2
