--- src/web/templates/dashboard.html
@@ -1,163 +1,0 @@
-<html>
-  <head>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.9.3/Chart.min.js" integrity="sha256-R4pqcOYV8lt7snxMQO/HSbVCFRPMdrhAFMH+vr9giYI=" crossorigin="anonymous"></script>
-  </head>
-  <body>
-    <div>
-      <div style="height: 400px; width: 800px">
-        <canvas id="myChart"></canvas>
-      </div>
-      <div>
-        {{ data.db }}
-      </div>
-      <div style="height: 400px; width: 800px">
-        <canvas id="myChart2"></canvas>
-      </div>
-      <script>
-        var ctx = document.getElementById('myChart');
-        var test = {{ data.alldb | tojson }}
-        let labels = test.map(item => item.db_size_timestamp)
-        let data = test.map(item => item.value_mb)
-
-        var myChart = new Chart(ctx, {
-        type: 'line',
-        data: {
-          labels: labels,
-          datasets: [{
-            label: 'Size in MB',
-            data: data,
-            backgroundColor: [
-                'rgba(255, 99, 132, 0.2)',
-                'rgba(54, 162, 235, 0.2)',
-                'rgba(255, 206, 86, 0.2)',
-                'rgba(75, 192, 192, 0.2)',
-                'rgba(153, 102, 255, 0.2)',
-                'rgba(255, 159, 64, 0.2)'
-            ],
-            borderColor: [
-                'rgba(255, 99, 132, 1)',
-                'rgba(54, 162, 235, 1)',
-                'rgba(255, 206, 86, 1)',
-                'rgba(75, 192, 192, 1)',
-                'rgba(153, 102, 255, 1)',
-                'rgba(255, 159, 64, 1)'
-            ],
-            borderWidth: 1
-            }]
-          },      
-          options: {
-              scales: {
-                  yAxes: [{
-                      ticks: {
-                        beginAtZero: true
-                          }
-                      }]
-                  }
-              }
-          });
-      </script>
-      <script>
-      var ctx2 = document.getElementById('myChart2');
-      var karma = {{ data.allkarma | tojson }}
-      let karmalabels = karma.map(item => item.karma_timestamp)
-      let comment_karma = karma.map(item => item.value.comment_karma)
-
-      let link_karma = karma.map(item => item.value.link_karma)
-      let total = karma.map(item => item.value.link_karma + item.value.comment_karma)
-
-        var myChart2 = new Chart(ctx2, {
-        type: 'line',
-        data: {
-          labels: karmalabels,
-          datasets: [
-          {
-            label: "Comment Karma",
-            backgroundColor: [
-                'rgba(0, 99, 132, 0.2)',
-                'rgba(54, 162, 235, 0.2)',
-                'rgba(255, 206, 86, 0.2)',
-                'rgba(75, 192, 192, 0.2)',
-                'rgba(153, 102, 255, 0.2)',
-                'rgba(255, 159, 64, 0.2)'
-            ],
-            borderColor: [
-                'rgba(255, 99, 132, 1)',
-                'rgba(54, 162, 235, 1)',
-                'rgba(255, 206, 86, 1)',
-                'rgba(75, 192, 192, 1)',
-                'rgba(153, 102, 255, 1)',
-                'rgba(255, 159, 64, 1)'
-            ],
-            borderWidth: 1,
-            // notice the gap in the data and the spanGaps: true
-            data: comment_karma,
-            fill: false,
-            spanGaps: true,
-          },
-          {
-            label: "Post Karma",
-            backgroundColor: [
-                'rgba(255, 99, 132, 0.2)',
-                'rgba(54, 162, 235, 0.2)',
-                'rgba(255, 206, 86, 0.2)',
-                'rgba(75, 192, 192, 0.2)',
-                'rgba(153, 102, 255, 0.2)',
-                'rgba(255, 159, 64, 0.2)'
-            ],
-            borderColor: [
-                'rgba(50, 99, 132, 1)',
-                'rgba(54, 162, 235, 1)',
-                'rgba(255, 206, 86, 1)',
-                'rgba(75, 192, 192, 1)',
-                'rgba(153, 102, 255, 1)',
-                'rgba(255, 159, 64, 1)'
-            ],
-            borderWidth: 1,
-            // notice the gap in the data and the spanGaps: true
-            data: link_karma,
-            spanGaps: true,
-            fill: false
-          },
-          {
-            label: "Total Karma",
-            backgroundColor: [
-                'rgba(2, 5, 34, 0.2)',
-                'rgba(54, 162, 235, 0.2)',
-                'rgba(255, 206, 86, 0.2)',
-                'rgba(75, 192, 192, 0.2)',
-                'rgba(153, 102, 255, 0.2)',
-                'rgba(255, 159, 64, 0.2)'
-            ],
-            borderColor: [
-                'rgba(50, 0, 132, 1)',
-                'rgba(54, 162, 235, 1)',
-                'rgba(255, 206, 86, 1)',
-                'rgba(75, 192, 192, 1)',
-                'rgba(153, 102, 255, 1)',
-                'rgba(255, 159, 64, 1)'
-            ],
-            borderWidth: 1,
-            // notice the gap in the data and the spanGaps: true
-            data: total,
-            fill: false,
-            spanGaps: true,
-          }
-        ]
-        },      
-        options: {
-            scales: {
-                yAxes: [{
-                    ticks: {
-                      beginAtZero: true
-                        }
-                    }]
-                }
-            }
-        });
-      </script>
-    </div>
-    <pre>
-      {{ data.user }}
-    </pre>
-  </body>
-</html>
